[refac] move exceptions
