+++ src/main/java/org/hankki/hankkiserver/api/auth/service/AuthService.java
@@ -8,9 +8,9 @@ import org.hankki.hankkiserver.common.code.ErrorCode;
 import org.hankki.hankkiserver.domain.user.model.User;
 import org.hankki.hankkiserver.domain.user.model.UserInfo;
 import org.hankki.hankkiserver.domain.user.model.Platform;
-import org.hankki.hankkiserver.exception.EntityNotFoundException;
-import org.hankki.hankkiserver.exception.InvalidValueException;
-import org.hankki.hankkiserver.exception.UnauthorizedException;
+import org.hankki.hankkiserver.common.exception.EntityNotFoundException;
+import org.hankki.hankkiserver.common.exception.InvalidValueException;
+import org.hankki.hankkiserver.common.exception.UnauthorizedException;
 import org.hankki.hankkiserver.oauth.apple.AppleOAuthProvider;
 import org.hankki.hankkiserver.oauth.dto.SocialInfoDto;
 import org.hankki.hankkiserver.oauth.kakao.KakaoOAuthProvider;
