+++ src/main/java/org/hankki/hankkiserver/oauth/apple/ApplePublicKeyGenerator.java
@@ -1,7 +1,7 @@
 package org.hankki.hankkiserver.oauth.apple;
 
 import org.hankki.hankkiserver.common.code.ErrorCode;
-import org.hankki.hankkiserver.exception.UnauthorizedException;
+import org.hankki.hankkiserver.common.exception.UnauthorizedException;
 import org.hankki.hankkiserver.oauth.apple.dto.ApplePublicKey;
 import org.hankki.hankkiserver.oauth.apple.dto.ApplePublicKeys;
 import org.springframework.stereotype.Component;
