+++ src/main/java/org/hankki/hankkiserver/oauth/apple/dto/ApplePublicKeys.java
@@ -1,7 +1,7 @@
 package org.hankki.hankkiserver.oauth.apple.dto;
 
 import org.hankki.hankkiserver.common.code.ErrorCode;
-import org.hankki.hankkiserver.exception.UnauthorizedException;
+import org.hankki.hankkiserver.common.exception.UnauthorizedException;
 
 import java.util.List;
